cmd: verbose mode from yaml
