+++ internal/cmd.go
@@ -1,28 +1,15 @@
 package internal
 
 import (
-	"flag"
 	"fmt"
 	"log/slog"
 	"os"
 
-	"github.com/fljdin/dispatch/internal/config"
 	"github.com/fljdin/dispatch/internal/dispatcher"
 	"github.com/lithammer/dedent"
 )
 
 var (
-	argConfig        string
-	argConfigDesc    string = "configuration file"
-	argOutput        string
-	argOutputDesc    string = "redirect output to file"
-	argProcesses     int
-	argProcessesDesc string = fmt.Sprintf("number of processes (default %d)", config.ProcessesDefault)
-	argVerbose       bool
-	argVerboseDesc   string = "verbose mode"
-	argVersion       bool
-	argVersionDesc   string = "show version"
-
 	usageStr string = dedent.Dedent(`
 		Usage:
 		  dispatch [options]
@@ -45,18 +32,10 @@ var (
 		argVersionDesc)
 )
 
-func newConfig() (config.Config, error) {
-	return config.NewBuilder().
-		FromYAML(argConfig).
-		WithProcesses(argProcesses).
-		WithLogfile(argOutput).
-		Build()
-}
-
 func Dispatch(version string) {
 	parseFlags()
 	setEnvirons()
-	setupLogging(os.Stderr)
+	setupLogging(os.Stderr, false)
 
 	if argVersion {
 		fmt.Println(version)
@@ -73,6 +52,8 @@ func Dispatch(version string) {
 		slog.Error(err.Error())
 		os.Exit(1)
 	}
+
+	setupLogging(os.Stderr, config.Verbose)
 
 	t, err := config.Tasks()
 	if err != nil {
@@ -91,7 +72,8 @@ func Dispatch(version string) {
 			slog.Error(err.Error())
 			os.Exit(1)
 		}
-		setupLogging(f)
+
+		setupLogging(f, config.Verbose)
 	}
 
 	dispatcher := dispatcher.New(config.Processes)
@@ -104,32 +86,11 @@ func Dispatch(version string) {
 		"loading configuration",
 		"tasks", len(t),
 		"procs", config.Processes,
+		"verbose", config.Verbose,
 	)
 
 	dispatcher.Wait()
 	os.Exit(0)
-}
-
-func parseFlags() {
-	flag.Usage = func() {
-		fmt.Fprint(flag.CommandLine.Output(), usage)
-	}
-
-	flag.StringVar(&argConfig, "c", "", argConfigDesc)
-	flag.StringVar(&argConfig, "config", "", argConfigDesc)
-
-	flag.StringVar(&argOutput, "o", "", argOutputDesc)
-	flag.StringVar(&argOutput, "output", "", argOutputDesc)
-
-	flag.IntVar(&argProcesses, "P", 0, argProcessesDesc)
-	flag.IntVar(&argProcesses, "procs", 0, argProcessesDesc)
-
-	flag.BoolVar(&argVerbose, "v", false, argVerboseDesc)
-	flag.BoolVar(&argVerbose, "verbose", false, argVerboseDesc)
-
-	flag.BoolVar(&argVersion, "version", false, argVersionDesc)
-
-	flag.Parse()
 }
 
 func setEnvirons() {
