+++ t/config/depends_on_loader_task.yaml
@@ -1,5 +1,6 @@
 # related to fljdin/dispatch#35
 procs: 2
+verbose: true
 logfile: depends_on_loader_task.log
 
 tasks:
