+++ t/expected/depends_on_loader_task.log
@@ -1,4 +1,4 @@
-2000-01-01 00:00:00 INFO   loading configuration tasks=2 procs=2
+2000-01-01 00:00:00 INFO   loading configuration tasks=2 procs=2 verbose=true
 2000-01-01 00:00:00 DEBUG  task=1:0 msg="task sent to internal channel"
 2000-01-01 00:00:00 INFO   task=1:0 status=succeeded name="task #1 must load two others" start="2000-01-01 00:00:00" elapsed=0s
 2000-01-01 00:00:00 INFO   task=1:0 cmd=sh action: echo -n "sleep .2\nsleep .1\n"
